Timing fixes
- change data types of runtime gathering to unsigend int
- fix primitive description
- make timing functions weak to be easier to redefine in application

The last point should maybe done to all these platform or application-dependent
functions.

+++ src/interpreter.c
@@ -636,7 +636,7 @@ void interpreter(short_jump_target start_base_address) {
 			/* end timer ( -- usecs secs ) */
 		case tend:
 			{
-				long int sec,usec;
+				unsigned int sec,usec;
 				end_timer(&sec,&usec);
 				ppush(usec);
 				ppush(sec);
