+++ src/runtime.h
@@ -1,3 +1,3 @@
 
 void start_timer(void);
-void end_timer(long int * sec, long int * usec);
+void end_timer(unsigned int * sec, unsigned int * usec);
